% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecommerce.R
\name{gmr_enhanced_ecom}
\alias{gmr_enhanced_ecom}
\title{Build Enhanced Ecommerce hits}
\usage{
gmr_enhanced_ecom(action = c("purchase", "detail", "click", "add", "remove",
  "checkout", "checkout_option", "purchase", "refund"), transaction_id,
  revenue = NULL, product_sku = NULL, product_name = NULL,
  product_brand = NULL, product_category = NULL, product_price = NULL,
  ...)
}
\arguments{
\item{action}{The role of the products included in a hit. Required.}

\item{transaction_id}{An id for the transaction. Required.}

\item{revenue}{The total value of the transaction}

\item{product_sku}{A vector of product SKU}

\item{product_name}{A vector of product names}

\item{product_brand}{A vector of product brands}

\item{product_category}{A vector of product category}

\item{product_price}{A vector of product price}
}
\value{
An \code{gmr_ee} object that can be passed to \code{gmr_hit_x} functions.
}
\description{
A helper function to build enhanced ecommerce hits
}
\details{
This lets you more easily create enhanced ecommerce hits by
 creating an object you then pass into other hits such as \link{gmr_hit_event}

The product arguments such as \link{product_sku} are vectors that will be turned into lists via \link{gmr_enhanced_index}.  If passing multiple arguments, they each must have the products in the same order and be the same length.

If you want to pass your own indexed products, \link{gmr_enhanced_index} can be used to create a named list which you can then append to the paramters generated using \link{do.call}.  See examples.
}
\examples{

gmr_hit_page(url_path = "/checkout",
             enhanced_ecom = gmr_enhanced_ecom("purchase", transaction_id = "blah223", revenue = 4300.23, product_sku = c("sku4","sku23","sku7")))

# Make your own hit
my_promotion_id <- gmr_enhanced_index(c("aff3","aff2","aff2"), prefix = "promo", suffix = "id")

# make your own enhanced ecom obj
my_ee <- do.call(gmr_enhanced_ecom,
                 args = c(list(action = "purchase",
                              transaction_id = "2323",
                              product_sku = c("sku4","sku3","sku7")
                              ),
                         my_promotion_id)
                         )

# register hit
gmr_hit_page(url_path = "/checkout-thanks", enhanced_ecom = my_ee)

}
