% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecommerce.R
\name{gmr_enhanced_index}
\alias{gmr_enhanced_index}
\title{Create ecommerce index}
\usage{
gmr_enhanced_index(my_vector = NULL, prefix = NULL, suffix = NULL)
}
\arguments{
\item{prefix}{The parameter prefix}

\item{suffix}{The parameter suffix

Pass a list of items you want to create, they will be indexed in the same order as the vector.}

\item{vector}{A vector of entries to send}
}
\description{
Enhanced e-commerce calls for product index's -
  this helps create them to pass to other calls such as \link{gmr_enhanced_ecom}
}
\examples{

gmr_enhanced_index(c("product1", "product2", "product3"), prefix = "pr", suffix = "id")

}
