% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_types.R
\name{gmr_hit_event}
\alias{gmr_hit_event}
\title{Send a event type hit}
\usage{
gmr_hit_event(category, action, label = NULL, value = NULL,
  interaction = TRUE, enhanced_ecom = NULL, ...)
}
\arguments{
\item{category}{Specifies the event category}

\item{action}{Specifies the event action.}

\item{label}{Optional label}

\item{value}{Optional value}

\item{interaction}{Whether hit is interactive with page (affecting bounce rate)}

\item{enhanced_ecom}{A \code{gmr_ee} generated via \link{gmr_enhanced_ecom}}

\item{...}{Other named arguments to pass in as payload data}
}
\description{
Suitable for tracking general events
}
\details{
Helper wrapper to send event type hits.  Specify named arguments to also send those, including overriding the helper arguments.  E.g. set your own cid = "fixed-value" if you want to have hits to be recorded as coming from the same user.
}
\examples{

gmr_hit_event("R_cat","R_action","R_label",300)


}
\seealso{
\url{https://developers.google.com/analytics/devguides/collection/protocol/v1/parameters}
}
