% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_types.R
\name{gmr_hit_page}
\alias{gmr_hit_page}
\title{Send a page type hit}
\usage{
gmr_hit_page(url_path = NULL, url_host = NULL, url_title = NULL, ...)
}
\arguments{
\item{url_path}{The path portion of the page URL, should start with \code{/}}

\item{url_host}{Specifies the hostname from which content was hosted}

\item{url_title}{The title of the page or document.}

\item{...}{Other named arguments to pass in as payload data}
}
\description{
Suitable for tracking pages
}
\details{
Helper wrapper to send page type hits.  Specify named arguments to also send those, including overriding the helper arguments.  E.g. set your own cid = "fixed-value" if you want to have hits to be recorded as coming from the same user.
}
\examples{

# send hit as a page
gmr_hit_page("/hi-from-r-example")

# override the dp argument (page path)
gmr_hit_page("/hi-from-r-test", dp = "/bye-from-r-test")

# set other arguments
my_cid <- gmr_uuid()
gmr_hit_page("/hi-from-r-test", cid = my_cid)

}
\seealso{
\url{https://developers.google.com/analytics/devguides/collection/protocol/v1/parameters}
}
