% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_types.R
\name{gmr_hit_timing}
\alias{gmr_hit_timing}
\title{Send a timing type hit}
\usage{
gmr_hit_timing(category, name, time, label = NULL, enhanced_ecom = NULL,
  ...)
}
\arguments{
\item{category}{Specifies the timing category}

\item{name}{Specifies the timing name}

\item{time}{An integer in milliseconds}

\item{label}{Optional label}

\item{enhanced_ecom}{A \code{gmr_ee} generated via \link{gmr_enhanced_ecom}}

\item{...}{Other named arguments to pass in as payload data}
}
\description{
Suitable for tracking time lags
}
\details{
Helper wrapper to send timing type hits.  Specify named arguments to also send those, including overriding the helper arguments.  E.g. set your own cid = "fixed-value" if you want to have hits to be recorded as coming from the same user.
}
\examples{

gmr_hit_timing("R_app","shiny_loadtime",3000, label = "appname")


}
\seealso{
\url{https://developers.google.com/analytics/devguides/collection/protocol/v1/parameters}
}
