% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_call.R
\name{gmr_post}
\alias{gmr_post}
\title{Post data to Measurement Protocol}
\usage{
gmr_post(payload_data, tid = Sys.getenv("MP_TRACKING_ID"),
  hittype = c("pageview", "screenview", "event", "transaction", "item",
  "social", "exception", "timing"))
}
\arguments{
\item{payload_data}{A named list of paramters to send.}

\item{tid}{Tracking ID.  By default from the Sys environment \code{MP_TRACKING_ID}, overwritten if tid also sent in \code{payload_data}}

\item{hittype}{Either specify in argument \code{payload_data$t} or here.}
}
\value{
TRUE if successful
}
\description{
Send data to the Google Analytics API endpoint.
}
\details{
Hits are sent via \code{httr::POST} and are URL encoded.

The version argument is set for you:  \code{v=1L}

\code{ip} is implicitly sent in the HTTP request and is used to compute all the geo / network dimensions in Google Analytics.  Override with \code{payload_data$uip = 1.2.3.4}

Set options("googleMeasureR.debug" = TRUE) to activate debug mode.
}
\seealso{
\url{https://developers.google.com/analytics/devguides/collection/protocol/v1/}
}
